correct for some warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21778 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/jp2kak/subfile_source.h
@@ -55,8 +55,8 @@ class subfile_source : public kdu_compressed_source {
               char** papszTokens = CSLTokenizeString2(fname + 12, ",", 0);
               if (CSLCount(papszTokens) >= 2)
               {
-                  subfile_offset = CPLScanUIntBig(papszTokens[0], strlen(papszTokens[0]));
-                  subfile_size = CPLScanUIntBig(papszTokens[1], strlen(papszTokens[1]));
+                  subfile_offset = (int) CPLScanUIntBig(papszTokens[0], strlen(papszTokens[0]));
+                  subfile_size = (int) CPLScanUIntBig(papszTokens[1], strlen(papszTokens[1]));
               }
               else
               {
